@@ -1,9 +1,9 @@
-﻿using MediatR;
+﻿using System.Text;
+using System.Text.Json;
+using MediatR;
 using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
-using System.Text;
 
 namespace Core.Application.Pipelines.Caching;
 
@@ -11,20 +11,20 @@ public class CachingBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, 
     where TRequest : IRequest<TResponse>, ICachableRequest
 {
     private readonly IDistributedCache _cache;
-    private readonly ILogger<CachingBehavior<TRequest, TResponse>> _logger;
 
     private readonly CacheSettings _cacheSettings;
+    private readonly ILogger<CachingBehavior<TRequest, TResponse>> _logger;
 
     public CachingBehavior(IDistributedCache cache, ILogger<CachingBehavior<TRequest, TResponse>> logger,
-                           IConfiguration configuration)
+        IConfiguration configuration)
     {
         _cache = cache;
         _logger = logger;
         _cacheSettings = configuration.GetSection("CacheSettings").Get<CacheSettings>();
     }
 
-    public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken,
-                                        RequestHandlerDelegate<TResponse> next)
+    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next,
+        CancellationToken cancellationToken)
     {
         TResponse response;
         if (request.BypassCache) return await next();
@@ -35,7 +35,7 @@ public class CachingBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, 
             TimeSpan? slidingExpiration =
                 request.SlidingExpiration ?? TimeSpan.FromDays(_cacheSettings.SlidingExpiration);
             DistributedCacheEntryOptions cacheOptions = new() { SlidingExpiration = slidingExpiration };
-            byte[] serializeData = Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(response));
+            byte[] serializeData = Encoding.UTF8.GetBytes(JsonSerializer.Serialize(response));
             await _cache.SetAsync(request.CacheKey, serializeData, cacheOptions, cancellationToken);
             return response;
         }
@@ -43,7 +43,7 @@ public class CachingBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, 
         byte[]? cachedResponse = await _cache.GetAsync(request.CacheKey, cancellationToken);
         if (cachedResponse != null)
         {
-            response = JsonConvert.DeserializeObject<TResponse>(Encoding.Default.GetString(cachedResponse));
+            response = JsonSerializer.Deserialize<TResponse>(Encoding.Default.GetString(cachedResponse));
             _logger.LogInformation($"Fetched from Cache -> {request.CacheKey}");
         }
         else
