@@ -1,16 +1,16 @@
-﻿using Core.CrossCuttingConcerns.Logging;
+﻿using System.Text.Json;
+using Core.CrossCuttingConcerns.Logging;
 using Core.CrossCuttingConcerns.Logging.Serilog;
 using MediatR;
 using Microsoft.AspNetCore.Http;
-using Newtonsoft.Json;
 
 namespace Core.Application.Pipelines.Logging;
 
 public class LoggingBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
     where TRequest : IRequest<TResponse>, ILoggableRequest
 {
-    private readonly LoggerServiceBase _loggerServiceBase;
     private readonly IHttpContextAccessor _httpContextAccessor;
+    private readonly LoggerServiceBase _loggerServiceBase;
 
 
     public LoggingBehavior(LoggerServiceBase loggerServiceBase, IHttpContextAccessor httpContextAccessor)
@@ -19,8 +19,8 @@ public class LoggingBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, 
         _httpContextAccessor = httpContextAccessor;
     }
 
-    public Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken,
-                                  RequestHandlerDelegate<TResponse> next)
+    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next,
+        CancellationToken cancellationToken)
     {
         List<LogParameter> logParameters = new();
         logParameters.Add(new LogParameter
@@ -35,12 +35,11 @@ public class LoggingBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, 
             Parameters = logParameters,
             User = _httpContextAccessor.HttpContext == null ||
                    _httpContextAccessor.HttpContext.User.Identity.Name == null
-                       ? "?"
-                       : _httpContextAccessor.HttpContext.User.Identity.Name
+                ? "?"
+                : _httpContextAccessor.HttpContext.User.Identity.Name
         };
 
-        _loggerServiceBase.Info(JsonConvert.SerializeObject(logDetail));
-
-        return next();
+        _loggerServiceBase.Info(JsonSerializer.Serialize(logDetail));
+        return await next();
     }
 }
