@@ -1,29 +1,29 @@
-﻿using MediatR;
-using System.Transactions;
+﻿using System.Transactions;
+using MediatR;
 
-namespace Core.Application.Pipelines.Transaction
-{
-    public class TransactionScopeBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
+namespace Core.Application.Pipelines.Transaction;
+
+public class TransactionScopeBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
     where TRequest : IRequest<TResponse>
+{
+    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next,
+        CancellationToken cancellationToken)
     {
-        public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken, RequestHandlerDelegate<TResponse> next)
+        using (TransactionScope transactionScope = new(TransactionScopeAsyncFlowOption.Enabled))
         {
-            using (TransactionScope transactionScope = new(TransactionScopeAsyncFlowOption.Enabled))
+            TResponse response;
+            try
             {
-                TResponse response;
-                try
-                {
-                    response = await next();
-                    transactionScope.Complete();
-                }
-                catch (Exception)
-                {
-                    transactionScope.Dispose();
-                    throw;
-                }
-
-                return response;
+                response = await next();
+                transactionScope.Complete();
             }
+            catch (Exception)
+            {
+                transactionScope.Dispose();
+                throw;
+            }
+
+            return response;
         }
     }
 }
