@@ -1,19 +1,20 @@
-﻿using Core.CrossCuttingConcerns.Exceptions.Handlers;
+﻿using System.Text.Json;
+using Core.CrossCuttingConcerns.Exceptions.Handlers;
 using Core.CrossCuttingConcerns.Logging;
 using Core.CrossCuttingConcerns.Logging.Serilog;
 using Microsoft.AspNetCore.Http;
-using Newtonsoft.Json;
 
 namespace Core.CrossCuttingConcerns.Exceptions;
 
 public class ExceptionMiddleware
 {
-    private readonly RequestDelegate _next;
-    private readonly HttpExceptionHandler _httpExceptionHandler;
     private readonly IHttpContextAccessor _contextAccessor;
+    private readonly HttpExceptionHandler _httpExceptionHandler;
     private readonly LoggerServiceBase _loggerService;
+    private readonly RequestDelegate _next;
 
-    public ExceptionMiddleware(RequestDelegate next, IHttpContextAccessor contextAccessor, LoggerServiceBase loggerService)
+    public ExceptionMiddleware(RequestDelegate next, IHttpContextAccessor contextAccessor,
+        LoggerServiceBase loggerService)
     {
         _next = next;
         _contextAccessor = contextAccessor;
@@ -59,7 +60,7 @@ public class ExceptionMiddleware
             User = _contextAccessor.HttpContext?.User.Identity?.Name ?? "?"
         };
 
-        _loggerService.Info(JsonConvert.SerializeObject(logDetail));
+        _loggerService.Info(JsonSerializer.Serialize(logDetail));
         return Task.CompletedTask;
     }
 }
