@@ -1,6 +1,6 @@
-﻿using MediatR;
+﻿using System.Diagnostics;
+using MediatR;
 using Microsoft.Extensions.Logging;
-using System.Diagnostics;
 
 namespace Core.Application.Pipelines.Performance;
 
@@ -16,8 +16,8 @@ public class PerformanceBehavior<TRequest, TResponse> : IPipelineBehavior<TReque
         _stopwatch = stopwatch;
     }
 
-    public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken,
-                                        RequestHandlerDelegate<TResponse> next)
+    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next,
+        CancellationToken cancellationToken)
     {
         string requestName = request.GetType().Name;
 
@@ -44,4 +44,3 @@ public class PerformanceBehavior<TRequest, TResponse> : IPipelineBehavior<TReque
         return response;
     }
 }
-
