@@ -1,6 +1,7 @@
-﻿using FluentValidation;
-using FluentValidation.Results;
+﻿using Core.CrossCuttingConcerns.Exceptions.Types;
+using FluentValidation;
 using MediatR;
+using ValidationException = Core.CrossCuttingConcerns.Exceptions.Types.ValidationException;
 
 namespace Core.Application.Pipelines.Validation;
 
@@ -14,16 +15,23 @@ public class RequestValidationBehavior<TRequest, TResponse> : IPipelineBehavior<
         _validators = validators;
     }
 
-    public Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken,
-                                  RequestHandlerDelegate<TResponse> next)
+    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next,
+        CancellationToken cancellationToken)
     {
         ValidationContext<object> context = new(request);
-        List<ValidationFailure> failures = _validators
-                                           .Select(validator => validator.Validate(context))
-                                           .SelectMany(result => result.Errors)
-                                           .Where(failure => failure != null)
-                                           .ToList();
-        if (failures.Count != 0) throw new ValidationException(failures);
-        return next();
+        IEnumerable<ValidationExceptionModel> errors = _validators
+            .Select(validator => validator.Validate(context))
+            .SelectMany(result => result.Errors)
+            .Where(failure => failure != null)
+            .GroupBy(p => p.PropertyName, (propertyName, errors) => new ValidationExceptionModel
+            {
+                Property = propertyName,
+                Errors = errors.Select(e => e.ErrorMessage)
+            })
+            .ToList();
+
+        if (errors.Any()) throw new ValidationException(errors);
+        TResponse response = await next();
+        return response;
     }
 }
