@@ -8,9 +8,8 @@ internal class InternalServerErrorProblemDetails : ProblemDetails
     public InternalServerErrorProblemDetails(string detail)
     {
         Title = "Internal server error";
-        Detail = detail;
+        Detail = "Internal server error";
         Status = StatusCodes.Status500InternalServerError;
         Type = "https://example.com/probs/internal";
-        Instance = "";
     }
 }
