@@ -1,4 +1,6 @@
-﻿using Application.Services.AdditionalServiceService;
+﻿using System.Reflection;
+using Application.Services.AdditionalServiceService;
+using Application.Services.AuthenticatorService;
 using Application.Services.AuthService;
 using Application.Services.CarService;
 using Application.Services.CustomerService;
@@ -22,7 +24,6 @@ using Core.Mailing.MailKitImplementations;
 using FluentValidation;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
-using System.Reflection;
 
 namespace Application;
 
@@ -31,7 +32,8 @@ public static class ApplicationServiceRegistration
     public static IServiceCollection AddApplicationServices(this IServiceCollection services)
     {
         services.AddAutoMapper(Assembly.GetExecutingAssembly());
-        services.AddMediatR(Assembly.GetExecutingAssembly());
+        services.AddMediatR(
+            configuration => configuration.RegisterServicesFromAssembly(Assembly.GetExecutingAssembly()));
 
         services.AddSubClassesOfType(Assembly.GetExecutingAssembly(), typeof(BaseBusinessRules));
 
@@ -46,6 +48,7 @@ public static class ApplicationServiceRegistration
 
         services.AddScoped<IAdditionalServiceService, AdditionalServiceManager>();
         services.AddScoped<IAuthService, AuthManager>();
+        services.AddScoped<IAuthenticatorService, AuthenticatorManager>();
         services.AddScoped<ICarService, CarManager>();
         services.AddScoped<ICustomerService, CustomerManager>();
         services.AddScoped<IFindeksCreditRateService, FindeksCreditRateManager>();
@@ -62,24 +65,13 @@ public static class ApplicationServiceRegistration
         return services;
     }
 
-    public static IServiceCollection AddSubClassesOfType(
-        this IServiceCollection services,
-        Assembly assembly,
-        Type type,
+    public static IServiceCollection AddSubClassesOfType(this IServiceCollection services, Assembly assembly, Type type,
         Func<IServiceCollection, Type, IServiceCollection>? addWithLifeCycle = null)
     {
         var types = assembly.GetTypes().Where(t => t.IsSubclassOf(type) && type != t).ToList();
         foreach (var item in types)
-        {
-            if (addWithLifeCycle == null)
-            {
-                services.AddScoped(item);
-            }
-            else
-            {
-                addWithLifeCycle(services, type);
-            }
-        }
+            if (addWithLifeCycle == null) services.AddScoped(item);
+            else addWithLifeCycle(services, type);
         return services;
     }
 }
