@@ -1,5 +1,6 @@
-﻿using Application.Features.Auth.Rules;
-using Application.Services.AuthService;
+﻿using System.Web;
+using Application.Features.Auth.Rules;
+using Application.Services.AuthenticatorService;
 using Application.Services.Repositories;
 using Application.Services.UserService;
 using Core.Mailing;
@@ -7,7 +8,6 @@ using Core.Security.Entities;
 using Core.Security.Enums;
 using MediatR;
 using MimeKit;
-using System.Web;
 
 namespace Application.Features.Auth.Commands.EnableEmailAuthenticator;
 
@@ -18,24 +18,24 @@ public class EnableEmailAuthenticatorCommand : IRequest
 
     public class EnableEmailAuthenticatorCommandHandler : IRequestHandler<EnableEmailAuthenticatorCommand>
     {
-        private readonly IUserService _userService;
-        private readonly IAuthService _authService;
+        private readonly AuthBusinessRules _authBusinessRules;
+        private readonly IAuthenticatorService _authenticatorService;
         private readonly IEmailAuthenticatorRepository _emailAuthenticatorRepository;
         private readonly IMailService _mailService;
-        private readonly AuthBusinessRules _authBusinessRules;
+        private readonly IUserService _userService;
 
-        public EnableEmailAuthenticatorCommandHandler(IUserService userService, IAuthService authService,
-                                                      IEmailAuthenticatorRepository emailAuthenticatorRepository,
-                                                      IMailService mailService, AuthBusinessRules authBusinessRules)
+        public EnableEmailAuthenticatorCommandHandler(IUserService userService,
+            IEmailAuthenticatorRepository emailAuthenticatorRepository, IMailService mailService,
+            AuthBusinessRules authBusinessRules, IAuthenticatorService authenticatorService)
         {
             _userService = userService;
-            _authService = authService;
             _emailAuthenticatorRepository = emailAuthenticatorRepository;
             _mailService = mailService;
             _authBusinessRules = authBusinessRules;
+            _authenticatorService = authenticatorService;
         }
 
-        public async Task<Unit> Handle(EnableEmailAuthenticatorCommand request, CancellationToken cancellationToken)
+        public async Task Handle(EnableEmailAuthenticatorCommand request, CancellationToken cancellationToken)
         {
             User user = await _userService.GetById(request.UserId);
             await _authBusinessRules.UserShouldBeExists(user);
@@ -44,13 +44,13 @@ public class EnableEmailAuthenticatorCommand : IRequest
             user.AuthenticatorType = AuthenticatorType.Email;
             await _userService.Update(user);
 
-            EmailAuthenticator emailAuthenticator = await _authService.CreateEmailAuthenticator(user);
+            EmailAuthenticator emailAuthenticator = await _authenticatorService.CreateEmailAuthenticator(user);
             EmailAuthenticator addedEmailAuthenticator =
                 await _emailAuthenticatorRepository.AddAsync(emailAuthenticator);
 
             var toEmailList = new List<MailboxAddress>
             {
-                new($"{user.FirstName} {user.LastName}",user.Email)
+                new($"{user.FirstName} {user.LastName}", user.Email)
             };
 
             _mailService.SendMail(new Mail
@@ -60,8 +60,6 @@ public class EnableEmailAuthenticatorCommand : IRequest
                 TextBody =
                     $"Click on the link to verify your email: {request.VerifyEmailUrlPrefix}?ActivationKey={HttpUtility.UrlEncode(addedEmailAuthenticator.ActivationKey)}"
             });
-
-            return Unit.Value; //todo: return dto?
         }
     }
 }
