@@ -1,5 +1,5 @@
 ﻿using Application.Features.Auth.Rules;
-using Application.Services.AuthService;
+using Application.Services.AuthenticatorService;
 using Application.Services.Repositories;
 using Application.Services.UserService;
 using Core.Security.Entities;
@@ -15,23 +15,23 @@ public class EnableOtpAuthenticatorCommand : IRequest<EnabledOtpAuthenticatorRes
         EnableOtpAuthenticatorCommandHandler : IRequestHandler<EnableOtpAuthenticatorCommand,
             EnabledOtpAuthenticatorResponse>
     {
-        private readonly IUserService _userService;
-        private readonly IAuthService _authService;
-        private readonly IOtpAuthenticatorRepository _otpAuthenticatorRepository;
         private readonly AuthBusinessRules _authBusinessRules;
+        private readonly IAuthenticatorService _authenticatorService;
+        private readonly IOtpAuthenticatorRepository _otpAuthenticatorRepository;
+        private readonly IUserService _userService;
 
-        public EnableOtpAuthenticatorCommandHandler(IUserService userService, IAuthService authService,
-                                                    IOtpAuthenticatorRepository otpAuthenticatorRepository,
-                                                    AuthBusinessRules authBusinessRules)
+        public EnableOtpAuthenticatorCommandHandler(IUserService userService,
+            IOtpAuthenticatorRepository otpAuthenticatorRepository, AuthBusinessRules authBusinessRules,
+            IAuthenticatorService authenticatorService)
         {
             _userService = userService;
-            _authService = authService;
             _otpAuthenticatorRepository = otpAuthenticatorRepository;
             _authBusinessRules = authBusinessRules;
+            _authenticatorService = authenticatorService;
         }
 
         public async Task<EnabledOtpAuthenticatorResponse> Handle(EnableOtpAuthenticatorCommand request,
-                                                             CancellationToken cancellationToken)
+            CancellationToken cancellationToken)
         {
             User user = await _userService.GetById(request.UserId);
             await _authBusinessRules.UserShouldBeExists(user);
@@ -43,13 +43,13 @@ public class EnableOtpAuthenticatorCommand : IRequest<EnabledOtpAuthenticatorRes
             if (isExistsOtpAuthenticator is not null)
                 await _otpAuthenticatorRepository.DeleteAsync(isExistsOtpAuthenticator);
 
-            OtpAuthenticator newOtpAuthenticator = await _authService.CreateOtpAuthenticator(user);
+            OtpAuthenticator newOtpAuthenticator = await _authenticatorService.CreateOtpAuthenticator(user);
             OtpAuthenticator addedOtpAuthenticator =
                 await _otpAuthenticatorRepository.AddAsync(newOtpAuthenticator);
 
             EnabledOtpAuthenticatorResponse enabledOtpAuthenticatorDto = new()
             {
-                SecretKey = await _authService.ConvertSecretKeyToString(addedOtpAuthenticator.SecretKey)
+                SecretKey = await _authenticatorService.ConvertSecretKeyToString(addedOtpAuthenticator.SecretKey)
             };
             return enabledOtpAuthenticatorDto;
         }
