@@ -16,8 +16,8 @@ public class GetListByDynamicModelQuery : IRequest<GetListResponse<GetListByDyna
 
     public class GetListModelByDynamicQueryHandler : IRequestHandler<GetListByDynamicModelQuery, GetListResponse<GetListByDynamicModelListItemDto>>
     {
-        private readonly IModelRepository _modelRepository;
         private readonly IMapper _mapper;
+        private readonly IModelRepository _modelRepository;
 
         public GetListModelByDynamicQueryHandler(IModelRepository modelRepository, IMapper mapper)
         {
@@ -30,11 +30,11 @@ public class GetListByDynamicModelQuery : IRequest<GetListResponse<GetListByDyna
         {
             IPaginate<Model> models = await _modelRepository.GetListByDynamicAsync(
                                           request.DynamicQuery,
-                                          c => c.Include(c => c.Brand)
+                                          include: c => c.Include(c => c.Brand)
                                                 .Include(c => c.Fuel)
                                                 .Include(c => c.Transmission),
-                                          request.PageRequest.Page,
-                                          request.PageRequest.PageSize);
+                                          index: request.PageRequest.Page,
+                                          size: request.PageRequest.PageSize);
             var mappedModelListModel = _mapper.Map<GetListResponse<GetListByDynamicModelListItemDto>>(models);
             return mappedModelListModel;
         }
