@@ -1,7 +1,7 @@
 using Application.Features.OperationClaims.Constants;
 using Application.Services.Repositories;
 using Core.Application.Rules;
-using Core.CrossCuttingConcerns.Exceptions;
+using Core.CrossCuttingConcerns.Exceptions.Types;
 using Core.Security.Entities;
 
 namespace Application.Features.OperationClaims.Rules;
