@@ -13,12 +13,7 @@ public interface IAuthService
     public Task RevokeDescendantRefreshTokens(RefreshToken refreshToken, string ipAddress, string reason);
 
     public Task RevokeRefreshToken(RefreshToken token, string ipAddress, string? reason = null,
-                                   string? replacedByToken = null);
+        string? replacedByToken = null);
 
     public Task<RefreshToken> RotateRefreshToken(User user, RefreshToken refreshToken, string ipAddress);
-    public Task<EmailAuthenticator> CreateEmailAuthenticator(User user);
-    public Task<OtpAuthenticator> CreateOtpAuthenticator(User user);
-    public Task<string> ConvertSecretKeyToString(byte[] secretKey);
-    public Task SendAuthenticatorCode(User user);
-    public Task VerifyAuthenticatorCode(User user, string authenticatorCode);
 }
