@@ -10,11 +10,11 @@ namespace Persistence;
 public static class PersistenceServiceRegistration
 {
     public static IServiceCollection AddPersistenceServices(this IServiceCollection services,
-                                                            IConfiguration configuration)
+        IConfiguration configuration)
     {
         services.AddDbContext<BaseDbContext>(options =>
-                                                 options.UseSqlServer(
-                                                     configuration.GetConnectionString("RentACarConnectionString")));
+            options.UseSqlServer(
+                configuration.GetConnectionString("RentACarConnectionString")));
         services.AddScoped<IAdditionalServiceRepository, AdditionalServiceRepository>();
         services.AddScoped<IBrandRepository, BrandRepository>();
         services.AddScoped<ICarRepository, CarRepository>();
